Add support for canceled/failed integration_test pipeline
If a buildkite pipeline fails or is cancelled sufficient signals are NOT
passed to the script that was running within a docker container to
allow the script to gracefully clean up after itself.

As such there is a high potential for orphaned resources to be left
running in AWS if this happens to the integration_test pipeline.

This commit attempts to add expeditor listeners to cancellation or
failures of the integration_test pipeline and attempt to perform the
cleanup of orphaned resources associated with the aborted job.

Signed-off-by: Christopher A. Snapp <csnapp@chef.io>